feat: add documentLoader as an optional argument of check
